Split tool call transition families
