+++ crates/defra-agent/src/tool_call_lifecycle/transition/native.rs
@@ -1,0 +1,361 @@
+use super::*;
+
+impl ToolCallLifecycle {
+    /// Pending → Running. Creates the DefraDB row if missing; idempotent if
+    /// already in Running. Sets `started_at` to `now`.
+    pub async fn start_running(&mut self) -> Result<()> {
+        if self.state == ToolCallState::Running {
+            // Idempotent re-entry (retry path).
+            return Ok(());
+        }
+        self.ensure_state(&[ToolCallState::Pending], "start_running")?;
+
+        let now = Utc::now();
+        let started_at_str = now.to_rfc3339();
+        let deadline_at_str = self.deadline_at.to_rfc3339();
+        let escaped_request_id = escape_graphql_string(&self.request_id);
+        let escaped_session_id = escape_graphql_string(&self.session_id);
+        let escaped_tool_call_id = escape_graphql_string(&self.tool_call_id);
+        let escaped_tool_name = escape_graphql_string(&self.tool_name);
+        let escaped_args = escape_graphql_string(&self.args);
+        let tool_call_key = format!("{escaped_session_id}:{escaped_tool_call_id}");
+        let message_sequence = self.message_sequence;
+
+        // Persist bridge fields for both R4 subagent bridges and R6 ordinary
+        // backgrounded tool rows. Native foreground calls omit them for
+        // back-compat with older rows.
+        let bridge_fields = if self.is_bridge() {
+            let await_mode_str = self.await_mode.as_str();
+            let cancel_policy_str = self.cancel_policy.as_str();
+            let child_field = self
+                .child_request_id
+                .as_ref()
+                .map(|crid| {
+                    let escaped_crid = escape_graphql_string(crid);
+                    format!(r#"child_request_id: "{escaped_crid}","#)
+                })
+                .unwrap_or_default();
+            let unclaimed_deadline_field = self
+                .unclaimed_deadline_at
+                .map(|deadline| {
+                    let escaped_deadline = escape_graphql_string(
+                        &deadline.to_rfc3339_opts(chrono::SecondsFormat::Secs, true),
+                    );
+                    format!(r#"unclaimed_deadline_at: "{escaped_deadline}","#)
+                })
+                .unwrap_or_default();
+            format!(
+                r#"{child_field}
+                    {unclaimed_deadline_field}
+                    await_mode: "{await_mode_str}",
+                    cancel_policy: "{cancel_policy_str}","#
+            )
+        } else {
+            String::new()
+        };
+
+        let mutation = format!(
+            r#"mutation {{
+                create_AgentToolCall(input: {{
+                    tool_call_key: "{tool_call_key}",
+                    request_id: "{escaped_request_id}",
+                    session_id: "{escaped_session_id}",
+                    message_sequence: {message_sequence},
+                    tool_name: "{escaped_tool_name}",
+                    tool_call_id: "{escaped_tool_call_id}",
+                    args: "{escaped_args}",
+                    result: "",
+                    status: "called",
+                    lifecycle_state: "running",
+                    started_at: "{started_at_str}",
+                    deadline_at: "{deadline_at_str}",
+                    {bridge_fields}
+                    selected_service_id: null,
+                    selected_tool_name: null,
+                    tool_failure_class: null,
+                    latency_ms: null
+                }}) {{ _docID }}
+            }}"#
+        );
+
+        let resp = execute_mutation_with_retry(&self.node, &mutation, "start_running")
+            .await
+            .context("start_running mutation")?;
+
+        let doc_id = extract_doc_id_from_create_response(&resp)
+            .ok_or_else(|| anyhow!("create_AgentToolCall returned no _docID"))?;
+
+        self.doc_id = Some(doc_id);
+        self.state = ToolCallState::Running;
+        self.started_at = Some(now);
+        Ok(())
+    }
+
+    /// Running → Completed. Writes the tool result; sets completed_at,
+    /// latency_ms.
+    pub async fn complete(&mut self, result: &str) -> Result<()> {
+        self.ensure_state(&[ToolCallState::Running], "complete")?;
+        if self.is_bridge() {
+            return Err(IllegalToolCallTransition::NativeCompleteOnSubagentTool.into());
+        }
+
+        let doc_id = self
+            .doc_id
+            .as_ref()
+            .ok_or_else(|| anyhow!("complete called before start_running persisted a row"))?;
+        let now = Utc::now();
+        let started_at = self
+            .started_at
+            .ok_or_else(|| anyhow!("complete called without started_at set"))?;
+        let latency_ms = (now - started_at).num_milliseconds();
+
+        let escaped_result = escape_graphql_string(result);
+        let escaped_doc_id = escape_graphql_string(doc_id);
+        let now_str = now.to_rfc3339();
+        // DefraDB requires DateTime fields to be re-supplied on update to
+        // avoid a type-mismatch error when re-validating the document.
+        let started_at_str = started_at.to_rfc3339();
+        let deadline_at_str = self.deadline_at.to_rfc3339();
+        let unclaimed_deadline_clear = self.clear_unclaimed_deadline_fragment();
+
+        let mutation = format!(
+            r#"mutation {{
+                update_AgentToolCall(
+                    filter: {{ _docID: {{ _eq: "{escaped_doc_id}" }} }},
+                    input: {{
+                        result: "{escaped_result}",
+                        status: "completed",
+                        lifecycle_state: "completed",
+                        started_at: "{started_at_str}",
+                        deadline_at: "{deadline_at_str}",
+                        completed_at: "{now_str}",
+                        latency_ms: {latency_ms}
+                        {unclaimed_deadline_clear}
+                    }}
+                ) {{ _docID }}
+            }}"#
+        );
+
+        execute_mutation_with_retry(&self.node, &mutation, "complete")
+            .await
+            .context("complete mutation")?;
+
+        self.state = ToolCallState::Completed;
+        Ok(())
+    }
+
+    /// Running → Failed. For tool errors during execution. Sets failure_class.
+    pub async fn fail(&mut self, result: &str, failure: super::FailureClass) -> Result<()> {
+        self.ensure_state(&[ToolCallState::Running], "fail")?;
+        if self.is_bridge() {
+            return Err(IllegalToolCallTransition::NativeFailOnSubagentTool.into());
+        }
+
+        let doc_id = self
+            .doc_id
+            .as_ref()
+            .ok_or_else(|| anyhow!("fail called before start_running persisted a row"))?;
+        let now = Utc::now();
+        let started_at = self
+            .started_at
+            .ok_or_else(|| anyhow!("fail called without started_at set"))?;
+        let latency_ms = (now - started_at).num_milliseconds();
+
+        let escaped_result = escape_graphql_string(result);
+        let escaped_doc_id = escape_graphql_string(doc_id);
+        let now_str = now.to_rfc3339();
+        let failure_class_str = failure.as_str();
+        // DefraDB requires DateTime fields to be re-supplied on update.
+        let started_at_str = started_at.to_rfc3339();
+        let deadline_at_str = self.deadline_at.to_rfc3339();
+        let unclaimed_deadline_clear = self.clear_unclaimed_deadline_fragment();
+
+        let mutation = format!(
+            r#"mutation {{
+                update_AgentToolCall(
+                    filter: {{ _docID: {{ _eq: "{escaped_doc_id}" }} }},
+                    input: {{
+                        result: "{escaped_result}",
+                        status: "completed",
+                        lifecycle_state: "failed",
+                        started_at: "{started_at_str}",
+                        deadline_at: "{deadline_at_str}",
+                        completed_at: "{now_str}",
+                        tool_failure_class: "{failure_class_str}",
+                        latency_ms: {latency_ms}
+                        {unclaimed_deadline_clear}
+                    }}
+                ) {{ _docID }}
+            }}"#
+        );
+
+        execute_mutation_with_retry(&self.node, &mutation, "fail")
+            .await
+            .context("fail mutation")?;
+
+        self.state = ToolCallState::Failed;
+        self.failure_class = Some(failure);
+        Ok(())
+    }
+
+    /// Pending → Failed. Used when the dispatcher cannot start the call
+    /// (MCP service unreachable, argument parse failure pre-spawn).
+    pub async fn spawn_failed(&mut self, failure: super::FailureClass, reason: &str) -> Result<()> {
+        self.ensure_state(&[ToolCallState::Pending], "spawn_failed")?;
+
+        // Pending means the row hasn't been created yet. We create it
+        // directly in Failed state.
+        let now = Utc::now();
+        let started_at_str = now.to_rfc3339();
+        let deadline_at_str = self.deadline_at.to_rfc3339();
+        let escaped_request_id = escape_graphql_string(&self.request_id);
+        let escaped_session_id = escape_graphql_string(&self.session_id);
+        let escaped_tool_call_id = escape_graphql_string(&self.tool_call_id);
+        let escaped_tool_name = escape_graphql_string(&self.tool_name);
+        let escaped_args = escape_graphql_string(&self.args);
+        let escaped_result = escape_graphql_string(reason);
+        let tool_call_key = format!("{escaped_session_id}:{escaped_tool_call_id}");
+        let message_sequence = self.message_sequence;
+        let failure_class_str = failure.as_str();
+
+        let mutation = format!(
+            r#"mutation {{
+                create_AgentToolCall(input: {{
+                    tool_call_key: "{tool_call_key}",
+                    request_id: "{escaped_request_id}",
+                    session_id: "{escaped_session_id}",
+                    message_sequence: {message_sequence},
+                    tool_name: "{escaped_tool_name}",
+                    tool_call_id: "{escaped_tool_call_id}",
+                    args: "{escaped_args}",
+                    result: "{escaped_result}",
+                    status: "completed",
+                    lifecycle_state: "failed",
+                    started_at: null,
+                    deadline_at: "{deadline_at_str}",
+                    completed_at: "{started_at_str}",
+                    tool_failure_class: "{failure_class_str}",
+                    latency_ms: 0
+                }}) {{ _docID }}
+            }}"#
+        );
+
+        let resp = execute_mutation_with_retry(&self.node, &mutation, "spawn_failed")
+            .await
+            .context("spawn_failed mutation")?;
+
+        let doc_id = extract_doc_id_from_create_response(&resp)
+            .ok_or_else(|| anyhow!("create_AgentToolCall returned no _docID"))?;
+
+        self.doc_id = Some(doc_id);
+        self.state = ToolCallState::Failed;
+        self.failure_class = Some(failure);
+        Ok(())
+    }
+
+    /// Running → TimedOut. Called by the runtime deadline wrapper and startup
+    /// recovery when a running tool call exceeds its effective deadline.
+    pub async fn timeout(&mut self) -> Result<()> {
+        self.ensure_state(&[ToolCallState::Running], "timeout")?;
+
+        let doc_id = self
+            .doc_id
+            .as_ref()
+            .ok_or_else(|| anyhow!("timeout called before start_running persisted a row"))?;
+        let now = Utc::now();
+        let started_at = self
+            .started_at
+            .ok_or_else(|| anyhow!("timeout called without started_at set"))?;
+        let latency_ms = (now - started_at).num_milliseconds();
+
+        let escaped_doc_id = escape_graphql_string(doc_id);
+        let escaped_result = escape_graphql_string(&format!(
+            "tool call deadline exceeded at {}",
+            self.deadline_at.to_rfc3339()
+        ));
+        let now_str = now.to_rfc3339();
+        // DefraDB requires DateTime fields to be re-supplied on update.
+        let started_at_str = started_at.to_rfc3339();
+        let deadline_at_str = self.deadline_at.to_rfc3339();
+        let failure_class = FailureClass::External.as_str();
+        let unclaimed_deadline_clear = self.clear_unclaimed_deadline_fragment();
+
+        let mutation = format!(
+            r#"mutation {{
+                update_AgentToolCall(
+                    filter: {{ _docID: {{ _eq: "{escaped_doc_id}" }} }},
+                    input: {{
+                        result: "{escaped_result}",
+                        status: "completed",
+                        lifecycle_state: "timedOut",
+                        tool_failure_class: "{failure_class}",
+                        started_at: "{started_at_str}",
+                        deadline_at: "{deadline_at_str}",
+                        completed_at: "{now_str}",
+                        latency_ms: {latency_ms}
+                        {unclaimed_deadline_clear}
+                    }}
+                ) {{ _docID }}
+            }}"#
+        );
+
+        execute_mutation_with_retry(&self.node, &mutation, "timeout")
+            .await
+            .context("timeout mutation")?;
+
+        self.state = ToolCallState::TimedOut;
+        self.failure_class = Some(FailureClass::External);
+        Ok(())
+    }
+
+    /// Pending → Cancelled. Used when a tool call is cancelled before
+    /// dispatch creates a running row.
+    pub async fn cancel_before_dispatch(&mut self) -> Result<()> {
+        self.ensure_state(&[ToolCallState::Pending], "cancel_before_dispatch")?;
+
+        // Pending: row may not exist yet. Create directly in Cancelled.
+        let now = Utc::now();
+        let started_at_str = now.to_rfc3339();
+        let deadline_at_str = self.deadline_at.to_rfc3339();
+        let escaped_request_id = escape_graphql_string(&self.request_id);
+        let escaped_session_id = escape_graphql_string(&self.session_id);
+        let escaped_tool_call_id = escape_graphql_string(&self.tool_call_id);
+        let escaped_tool_name = escape_graphql_string(&self.tool_name);
+        let escaped_args = escape_graphql_string(&self.args);
+        let tool_call_key = format!("{escaped_session_id}:{escaped_tool_call_id}");
+        let message_sequence = self.message_sequence;
+
+        let escaped_result = escape_graphql_string("tool call cancelled before dispatch");
+
+        let mutation = format!(
+            r#"mutation {{
+                create_AgentToolCall(input: {{
+                    tool_call_key: "{tool_call_key}",
+                    request_id: "{escaped_request_id}",
+                    session_id: "{escaped_session_id}",
+                    message_sequence: {message_sequence},
+                    tool_name: "{escaped_tool_name}",
+                    tool_call_id: "{escaped_tool_call_id}",
+                    args: "{escaped_args}",
+                    result: "{escaped_result}",
+                    status: "completed",
+                    lifecycle_state: "cancelled",
+                    started_at: null,
+                    deadline_at: "{deadline_at_str}",
+                    completed_at: "{started_at_str}",
+                    latency_ms: 0
+                }}) {{ _docID }}
+            }}"#
+        );
+
+        let resp = execute_mutation_with_retry(&self.node, &mutation, "cancel_before_dispatch")
+            .await
+            .context("cancel_before_dispatch mutation")?;
+        let doc_id = extract_doc_id_from_create_response(&resp)
+            .ok_or_else(|| anyhow!("create_AgentToolCall returned no _docID"))?;
+
+        self.doc_id = Some(doc_id);
+        self.state = ToolCallState::Cancelled;
+        Ok(())
+    }
+}
